# api/docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  domain.CreateStudentInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  domain.Session:
    properties:
      expiresAt:
        type: string
      refreshToken:
        type: string
    type: object
...
  domain.Student:
    properties:
      availableCourses:
        items:
          type: string
        type: array
      availableModules:
        items:
          type: string
        type: array
      availableOffers:
        items:
          type: string
        type: array
      blocked:
        type: boolean
      email:
        type: string
      id:
        type: string
      lastVisitAt:
        type: string
      name:
        type: string
      password:
        type: string
      registeredAt:
        type: string
      schoolId:
        type: string
      session:
        $ref: '#/definitions/domain.Session'
      verification:
        $ref: '#/definitions/domain.Verification'
    type: object
...
  domain.UpdateStudentInput:
    properties:
      blocked:
        type: boolean
      email:
        type: string
      name:
        type: string
      verified:
        type: boolean
    type: object
  domain.Verification:
    properties:
      code:
        type: string
      verified:
        type: boolean
    type: object
  v1.adminGetVideoResponse:
    properties:
      status:
        type: integer
      url:
        type: string
    type: object
...
  v1.dataResponse:
    properties:
      count:
        type: integer
      data:
        type: object
    type: object
...
  v1.manageOfferPermissionInput:
    properties:
      available:
        type: boolean
    type: object
...
  v1.response:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: REST API for UltraThreads App
  title: UltraThreads API
  version: "1.0"
paths:
...
  /admins/media/videos/{id}:
    get:
      consumes:
      - application/json
      description: get video by id
      parameters:
      - description: video id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.adminGetVideoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Get Video By ID
      tags:
      - admins-media
...
  /admins/students:
    get:
      consumes:
      - application/json
      description: admin get all students
      parameters:
      - description: skip
        in: query
        name: skip
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - description: verified
        in: query
        name: verified
        type: boolean
      - description: registerDateFrom
        in: query
        name: registerDateFrom
        type: string
      - description: registerDateTo
        in: query
        name: registerDateTo
        type: string
      - description: lastVisitDateFrom
        in: query
        name: lastVisitDateFrom
        type: string
      - description: registerDateTo
        in: query
        name: lastVisitDateTo
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.dataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Get Students
      tags:
      - admins-students
    post:
      consumes:
      - application/json
      description: admin create student
      parameters:
      - description: student info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.CreateStudentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Create Student
      tags:
      - admins-students
  /admins/students/{id}:
    delete:
      consumes:
      - application/json
      description: admin delete student
      parameters:
      - description: student id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Delete Student
      tags:
      - admins-students
    get:
      consumes:
      - application/json
      description: admin get student by id
      parameters:
      - description: student id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Get Student By ID
      tags:
      - admins-students
    put:
      consumes:
      - application/json
      description: admin update student
      parameters:
      - description: update info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateStudentInput'
      - description: student id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Update Student
      tags:
      - admins-students
  /admins/students/{id}/offers/{offerId}:
    patch:
      consumes:
      - application/json
      description: admin give student access to offer
      parameters:
      - description: permission type
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.manageOfferPermissionInput'
      - description: student id
        in: path
        name: id
        required: true
        type: string
      - description: offer id
        in: path
        name: offerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Give Student Access to Offer
      tags:
      - admins-students
...
securityDefinitions:
  AdminAuth:
    in: header
    name: Authorization
    type: apiKey
  StudentsAuth:
    in: header
    name: Authorization
    type: apiKey
  UsersAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"